requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.3.0
django-cors-headers==4.3.1
Django==5.1.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
reportlab==4.3.1
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
whitenoise==6.9.0
